+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.14 2016-??-??
+ - 2016-10-05 Fixed handling hidden files and filter quotation fixes.
  - 2016-10-03 Fixed searching for strings with ! in Sphinx.
  - 2016-09-30 Attachment name validation added.
  - 2016-09-30 Fixed escaping values passed to LDAP filters.